Fix -Wextra warnings in pallenelib
Now that pallenelib is compiled with the rest of the Lua interpreter we are compiling with -Wextra,
which has revealed a couple of warnings. This commit fixes them.

1. The inline cache now has type int instead of size_t. This reflects how Lua itself uses int for
   this purpose. (For example, see the various places that use the sizenode macro)

2. In the implementation of io.write, the L parameter should be marked as unused.

+++ pallene/coder.lua
@@ -1265,7 +1265,7 @@ gen_cmd["GetTable"] = function(self, cmd, _func)
 
     return util.render([[
         {
-            static size_t cache = UINT_MAX;
+            static int cache = UINT_MAX;
             TValue *slot = pallene_getstr($field_len, $tab, $key, &cache);
             ${get_slot}
         }
@@ -1289,7 +1289,7 @@ gen_cmd["SetTable"] = function(self, cmd, _func)
     return util.render([[
         {
             TValue keyv; ${init_keyv}
-            static size_t cache = UINT_MAX;
+            static int cache = UINT_MAX;
             TValue *slot = pallene_getstr($field_len, $tab, $key, &cache);
             if (PALLENE_UNLIKELY(isabstkey(slot))) {
                 TValue valv; ${init_valv}
